model/requirements.txt
google==3.0.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
grpcio-tools==1.38.0
pandas==1.1.5
protobuf==3.16.0
requests==2.25.1
threadpoolctl==2.2.0
urllib3==1.26.5
fastapi==0.83.0
uvicorn==0.16.0
simplejson==3.17.6
Jinja2==3.1.2
python-multipart==0.0.5
lightgbm
